requirements.txt
# Azab Discord Bot - Dependencies
# ================================
#
# Required packages for the Azab Discord bot built for discord.gg/syria
#
# Author: حَـــــنَّـــــا
# Server: discord.gg/syria

# Core Discord API Library
# =========================
# Discord.py 2.7.0 dev for latest features
git+https://github.com/Rapptz/discord.py.git@master#egg=discord.py

# Environment Configuration
# ==========================
python-dotenv>=1.0.0

# HTTP Client
# ============
aiohttp>=3.9.1

# AI Integration
# ===============
openai>=1.40.0

# System Monitoring
# ==================
psutil>=5.9.0
